55- created new component to resolve TypeScript component props type conflict.
