+++ src/pages/AuctionListingPage.tsx
@@ -1,15 +1,16 @@
 import { Row, Col } from "react-bootstrap";
 import Image from "../parts/Image";
-import { useState } from "react";
-import type Auction from "../interfaces/Auction";
 import InputPlaceBid from "../components/InputPlaceBid";
 import BidHistory from "../components/BidHistory";
 import type { Bid } from "../interfaces/Bid";
 import type { Info } from "../components/GetInformation";
 import { GetInformation } from "../components/GetInformation";
+import AuctionListingBidding from "../components/AuctionListingBidding";
 
-AboutPage.route = {
+AuctionListingPage.route = {
   path: "/listing",
+  index: 2,
+  menulabel: "Auction Listing Page",
 };
 
 //Imported things and copied things directly from auktioncard. This will be changed in future.
@@ -37,62 +38,13 @@ const sampleInfo: Info = {
   freight: "500 SEK",
 };
 
-export default function AboutPage({
-  id = sampleAuction.id,
-  title = sampleAuction.title,
-  currentBid = sampleAuction.currentBid,
-  endTime = sampleAuction.endTime,
-  favorited = sampleAuction.favorited,
-}: Auction = sampleAuction) {
-  const [isFavorited, setFavorited] = useState(favorited);
-  const currentTime = new Date();
-  let remainingTimeMessage = "";
-  // Time Difference in Milliseconds
-  const timeDifferenceMilli: number = endTime.getTime() - currentTime.getTime();
-
-  if (timeDifferenceMilli < 0) {
-    remainingTimeMessage = "Auction has finished";
-  } else {
-    const remainingSeconds = Math.floor(timeDifferenceMilli / 1000);
-    const remainingMinutes = Math.floor(remainingSeconds / 60);
-    const remainingHours = Math.floor(remainingMinutes / 60);
-    const remainingDays = Math.floor(remainingHours / 24);
-
-    // Only display the largest time unit remaining
-    if (remainingDays > 0) {
-      remainingTimeMessage = `${remainingDays} days`;
-    } else if (remainingHours > 0) {
-      remainingTimeMessage = `${remainingHours} hours`;
-    } else if (remainingMinutes > 0) {
-      remainingTimeMessage = `${remainingMinutes} minutes`;
-    } else remainingTimeMessage = `${remainingSeconds} seconds`;
-  }
-
-  function onFavorite() {
-    setFavorited((isFavorited) => !isFavorited);
-  }
+export default function AuctionListingPage() {
   return (
     <Row>
       <Col>
         <Image src="/images/products/3.jpg" alt="Here is a product" />
 
-        <div className="d-flex justify-content-between align-items-center">
-          <h2 className="mb-0">{title}</h2>
-          <span role="button" onClick={onFavorite} className="me-3">
-            <i
-              className={`bi bi-suit-heart${isFavorited ? "-fill" : ""} fs-1`}
-            ></i>
-          </span>
-        </div>
-
-        <div>
-          {/*Current bid. start bid, timeleft.  */}
-          <p className="mb-1">Start Bid: </p> {/*  */}
-          <p>Current bid: {currentBid} SEK</p>
-          <small className="text-muted">
-            Time left: {remainingTimeMessage}
-          </small>
-        </div>
+        <AuctionListingBidding auction={sampleAuction} />
 
         <div>
           {/* Add bid here */}
